Move classes to other packages
