+++ core/src/test/java/org/frankframework/configuration/ClassNameRewriterTest.java
@@ -8,6 +8,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import org.frankframework.configuration.filters.ClassNameRewriter;
 import org.frankframework.pipes.GetFromSessionPipe;
 import org.frankframework.pipes.JsonWellFormedCheckerPipe;
 import org.frankframework.pipes.PutInSessionPipe;
