+++ core/src/test/java/org/frankframework/configuration/util/ConfigurationUtilsTest.java
@@ -1,4 +1,4 @@
-package org.frankframework.configuration;
+package org.frankframework.configuration.util;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
