+++ nn-specials/src/main/java/org/frankframework/extensions/esb/EsbSoapWrapperPipe.java
@@ -20,8 +20,8 @@ import org.apache.commons.lang3.StringUtils;
 import lombok.Getter;
 
 import org.frankframework.configuration.ConfigurationException;
-import org.frankframework.configuration.ConfigurationUtils;
 import org.frankframework.configuration.ConfigurationWarnings;
+import org.frankframework.configuration.util.ConfigurationUtils;
 import org.frankframework.core.Adapter;
 import org.frankframework.core.DestinationValidator;
 import org.frankframework.core.IListener;
